impr: Replace assignment `hash-chain`s with improved metadata
Prior to this commit, HyperBEAM's `~scheduler@1.0` largely mimicked AO legacynet's `hash-chain` implementation. This commit replaces that key of assignments with a more useful and executable `base-hashpath`. This hashpath tracks the application of each assignment atop each prior assignment and the base process definition message. While fulfilling the role of the previous mechanism, this approach also enables another possibility: A recipient of any assignment is in possession of a single path that will allow them to execute a process to the point at which the given assignment is applied. This execution can be achieved without having to contact the scheduler or even make calls to the scheduler device for the process.